+++ test/engine_test.rb
@@ -26,8 +26,10 @@ class EngineTest < Minitest::Test
     refute_includes show, "notice"
     refute_includes show, "flash"
     refute_includes show, "FlatPack::Alert::Component"
-    assert_includes show, "recording_studio_page_nav_right"
-    assert_includes show, "recording_access_management_link"
+    refute_includes show, "recording_studio_page_nav_right"
+    refute_includes show, "recording_access_management_link"
+    refute_includes show, "Manage access"
+    assert_includes show, "photo"
   end
 
   def test_profile_authorization_uses_accessible_not_current_user_acl
@@ -37,7 +39,12 @@ class EngineTest < Minitest::Test
     admin = File.read(File.expand_path("../app/controllers/recording_studio_user/admin/users_controller.rb", __dir__))
 
     assert_includes profile, "RecordingStudio.enable_capability(:accessible, on: self)"
+    assert_includes profile, "RecordingStudio::Capabilities::Attachable.to"
+    assert_includes profile, 'allowed_content_types: ["image/*"]'
+    assert_includes profile, "enabled_attachment_kinds: %i[image]"
+    assert_includes profile, "max_file_count: 1"
     refute_includes people, "enable_capability(:accessible"
+    refute_includes people, "Capabilities::Attachable"
     assert_includes controller, "RecordingStudioAccessible.authorized?"
     refute_includes controller, "can_access?"
     refute_includes controller, "@user.update"