+++ test/install_generator_test.rb
@@ -37,8 +37,10 @@ class InstallGeneratorTest < Minitest::Test
     refute_includes generator, "invoke \"recording_studio_user:admin\""
     assert_includes generator, "as: :recording_studio_users"
     assert_includes generator, "recording_studio_user:migrations"
-    assert_includes generator, "Accessible on Profile"
+    assert_includes generator, "Accessible and"
+    assert_includes generator, "Attachable on Profile"
     assert_includes generator, "bootstrap_owner_access!"
+    assert_includes generator, "recording_studio_attachable:install"
     refute_includes generator, "access_management_authorizer"
     refute_includes generator, "AccessCreationContext"
   end